requirements.txt
requests==2.32.5
PyYAML==6.0.3
python-slugify==8.0.4
tqdm==4.67.3
